+++ frame/nomination-pools/benchmarking/src/mock.rs
@@ -25,7 +25,7 @@ use sp_runtime::{
 
 type AccountId = u128;
 type AccountIndex = u32;
-type BlockNumber = u64;
+type BlockNumber = u32;
 type Balance = u128;
 
 impl frame_system::Config for Runtime {
